POC A way to unify the checkout
